# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: omu-rest
  description: 'Laravel製RESTful APIサーバー。著者: [connect0459](https://github.com/connect0459)'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api'
    description: ローカルホストのAPIエンドポイント
  -
    url: 'https://example.studyfortwo.org/omu-rest/public/api'
    description: lolipopのAPIエンドポイント
paths:
...
  /books_stocks:
    get:
      tags:
        - books_stocks
      summary: 書籍在庫のリストを取得する
      operationId: 46c90e16b416e254ed79948864ae87aa
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookStock'
    post:
      tags:
        - books_stocks
      summary: 新しい書籍在庫を作成する
      operationId: ecc649eea287e26856100c11f96bb3b9
      requestBody:
        description: BookStockのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookStock'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/BookStock'
                  -
                    properties:
                      created_at:
                        type: string
                        format: date-time
                      updated_at:
                        type: string
                        format: date-time
                    type: object
  '/books_stocks/{id}':
    get:
      tags:
        - books_stocks
      summary: 特定のIDで書籍在庫を取得する
      operationId: d1c6d70f9a35791050b5da811f1abda1
      parameters:
        -
          name: id
          in: path
          description: 書籍在庫のID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookStock'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - books_stocks
      summary: 特定のIDで書籍在庫を更新する
      operationId: dc7eef146e782e14cddc076176ff04c4
      parameters:
        -
          name: id
          in: path
          description: 書籍在庫のID
          required: true
          schema:
            type: string
      requestBody:
        description: BookStockのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookStock'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/BookStock'
                  -
                    properties:
                      created_at:
                        type: string
                        format: date-time
                      updated_at:
                        type: string
                        format: date-time
                    type: object
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - books_stocks
      summary: 特定のIDで書籍在庫を削除する
      operationId: 45a5cfdaf4972cd73313133082ca0d26
      parameters:
        -
          name: id
          in: path
          description: 書籍在庫のID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
  /contacts:
    get:
      tags:
        - contacts
      summary: 連絡先のリストを取得する
      operationId: e670331f5da6af785062a857ec92def2
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
    post:
      tags:
        - contacts
      summary: 新しい連絡先を作成する
      operationId: 77b53a0b0cb97bf04eafa192e4cbac4e
      requestBody:
        description: 連絡先のデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/Contact'
  '/contacts/{id}':
    get:
      tags:
        - contacts
      summary: 特定のIDで連絡先を取得する
      operationId: 4c4e002507e3a53acc022b210aad20ac
      parameters:
        -
          name: id
          in: path
          description: 連絡先のID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - contacts
      summary: 特定のIDで連絡先を更新する
      operationId: adc814a54ff3a1fdc424293b19ffe8a7
      parameters:
        -
          name: id
          in: path
          description: 連絡先のID
          required: true
          schema:
            type: string
      requestBody:
        description: 連絡先のデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/Contact'
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - contacts
      summary: 特定のIDで連絡先を削除する
      operationId: bc8b0b91abd89481c321fcdb966dc355
      parameters:
        -
          name: id
          in: path
          description: 連絡先のID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
...
components:
  schemas:
...
    BookStock:
      description: 'BookStock Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        type_branch_id:
          description: 外部キー制約（types_branchesテーブル）
          type: integer
          format: int64
          example: '1'
        book_info_id:
          description: 外部キー制約（books_infoテーブル）
          type: integer
          format: int64
          example: '1'
        stock:
          description: 未販売の在庫数
          type: integer
          format: int64
          example: '4'
        order:
          description: 未処理の予約数
          type: integer
          format: int64
          example: '1'
        sold:
          description: 販売済みの在庫数
          type: integer
          format: int64
          example: '1'
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
    Contact:
      description: 'Contact Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        unique:
          description: お問い合わせ番号として相手に送信する一意な値
          type: string
          format: string
          example: 230C001
        name:
          description: 名前
          type: string
          format: string
          example: SFT太郎
        belong:
          description: 所属
          type: string
          format: string
          example: SFT大学
        email:
          description: メールアドレス
          type: string
          format: string
          example: example@studyfortwo.org
        summary:
          description: 概要
          type: string
          format: string
          example: メンバー加入について
        detail:
          description: 詳細
          type: string
          format: string
          example: 活動に興味があるのですが、この時期からでも入れますか？
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
...
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: api_token
      in: header